requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
google-genai
yfinance
requests
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
psycopg2-binary
pydantic[email]
email-validator
resend